Se migra register/login/logout/avatar a AppAuth
