+++ AppAuth/admin.py
@@ -1,3 +1,7 @@
 from django.contrib import admin
+from AppAuth.models import *
+
 
 # Register your models here.
+
+admin.site.register(Avatar)
