+++ AppAuth/templates/AppAuth/editarperfil.html
@@ -15,7 +15,11 @@
 
     <h1>Edicion de usuario</h1>
 
-    {{ form.as_p }}
+    <table>
+
+    {{ form.as_table }}
+
+    </table>
 
     <input type="submit" value="Guardar">
