+++ AppAuth/templates/AppAuth/login.html
@@ -14,7 +14,11 @@
 
     {% endif %}
 
-    {{form.as_p}}
+    <table>
+
+    {{form.as_table}}
+
+    </table>
 
     <input type="submit" value="Iniciar Sesion">
