+++ AppAuth/templates/AppAuth/register.html
@@ -13,7 +13,11 @@
 
     {% endif %}
 
-    {{form.as_p}}
+    <table>
+
+    {{form.as_table}}
+
+    </table>
 
     <input type="submit" value="Registrar Usuario">
