+++ AppLibreria/admin.py
@@ -1,11 +1,8 @@
 from django.contrib import admin
 from AppLibreria.models import *
 
-
-
 # Register your models here.
 
 admin.site.register(Libro)
 admin.site.register(Proveedores)
 admin.site.register(Cliente)
-admin.site.register(Avatar)
