+++ AppLibreria/urls.py
@@ -1,6 +1,7 @@
 from django.urls import path
 from AppLibreria.views import *
 from django.contrib.auth.views import LogoutView
+from AppAuth.views import *
 
 
 urlpatterns = [
@@ -12,10 +13,9 @@ urlpatterns = [
     path("proveedores/", proveedores, name="proveedores"),
     path("stock/", Leerstock, name="stock"),
     path("login/", iniciar_sesion, name="auth_login"),
-    path("logout/",LogoutView.as_view(template_name='AppLibreria/logout.html'), name="auth_logout"),
+    path("logout/",LogoutView.as_view(template_name='AppAuth/logout.html'), name="auth_logout"),
     path("register/", registrar_usuario, name="reg_user"),
     path("edit_user/", editar_perfil, name="auth_edit"),
     path("add_avatar/", agregar_avatar, name="auth_avatar"),
-
 
 ]
