+++ entrega1_vazquez/settings.py
@@ -39,6 +39,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'AppLibreria',
+    'AppAuth',
 ]
 
 MIDDLEWARE = [
